tar: avoid strtoul().
This is part of the general trend to prefer signed integer types,
to allow better runtime checking with -fsanitize=undefined etc.

  * (struct tar_stat_info): ue intmax_t, not unsigned, for sparse major
    and minor.
  * (tar_sparse_major): likewise.
  * (tar_sparse_minor): likewise.

Sync with GNU tar 4642cd04edbd57414e004920fa4976d9f3be6206.

+++ src/vfs/tar/tar-internal.h
@@ -268,8 +268,8 @@ struct tar_stat_info
     gboolean is_sparse;         /**< is the file sparse */
 
     /* For sparse files */
-    unsigned int sparse_major;
-    unsigned int sparse_minor;
+    intmax_t sparse_major;
+    intmax_t sparse_minor;
     GArray *sparse_map;         /**< array of struct sp_array */
 
     off_t real_size;            /**< real size of sparse file */
